marketplace/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
grpcio==1.57.0
grpcio-tools==1.57.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.24.1
Werkzeug==2.3.7
